Move `FeedCachePolicy` to its own file.
